transferred all necessary content to auth_helpers need to fix circular reference to db
