+++ hello.py
@@ -4,7 +4,7 @@ from flask_mail import Mail
 from datetime import datetime
 from passlib.hash import sha256_crypt
 from helpers import login_required
-from time import time
+from user_auth.routes import userAuth
 import jwt, os
 from dotenv import load_dotenv
 import requests
@@ -27,10 +27,14 @@ app.config['SESSION_TYPE'] = 'filesystem'
 mysql = MySQL(app)
 
 
+
+
 # Welcome page
 @app.route('/')
 def welcome():
     return render_template('welcome.html')
+
+
 
 # # Following Mailgun's standard api email format
 # def send_simple_message(subject, username, email, token, html, url, api, address):
@@ -114,264 +118,264 @@ def welcome():
 #         cur.close()
 #         return str(result)
 
-# Register user
-@app.route('/register', methods=['GET', 'POST'])
-def register():
-    if request.method == 'POST':
+# # Register user
+# @app.route('/register', methods=['GET', 'POST'])
+# def register():
+#     if request.method == 'POST':
 
-        # Gather user form data
-        name = request.form['name']
-        email = request.form['email']
-        username = request.form['username']
-        password = request.form['password']
-        confirm_password = request.form['confirm']
-        encrypt_password = sha256_crypt.hash(str(password))
-        dToday = datetime.now(tz=None).date()
+#         # Gather user form data
+#         name = request.form['name']
+#         email = request.form['email']
+#         username = request.form['username']
+#         password = request.form['password']
+#         confirm_password = request.form['confirm']
+#         encrypt_password = sha256_crypt.hash(str(password))
+#         dToday = datetime.now(tz=None).date()
 
-        # Make sure form data is acceptable
+#         # Make sure form data is acceptable
 
-        # Make sure no blank fields
-        if not name or not email or not username or not password:
-            message = "Registration fields cannot be blank"
-            return render_template('register.html', message=message)
+#         # Make sure no blank fields
+#         if not name or not email or not username or not password:
+#             message = "Registration fields cannot be blank"
+#             return render_template('register.html', message=message)
 
-        # Make sure email is unique
-        if check_email(email) != "0":
-            message = f"Email \"{str(email)}\" is already being used."
-            return render_template('register.html', message=message)
+#         # Make sure email is unique
+#         if check_email(email) != "0":
+#             message = f"Email \"{str(email)}\" is already being used."
+#             return render_template('register.html', message=message)
 
-        # Make sure username is >= 5 characters
-        if len(username) < 5:
-            message = "Username must be 5 characters minimum."
-            return render_template('register.html', message=message)
+#         # Make sure username is >= 5 characters
+#         if len(username) < 5:
+#             message = "Username must be 5 characters minimum."
+#             return render_template('register.html', message=message)
 
-        # Make sure username is unique
-        if verify_username(username) == "1":
-            message = f"Username \"{str(username)}\" already taken"
-            return render_template('register.html', message=message)
+#         # Make sure username is unique
+#         if verify_username(username) == "1":
+#             message = f"Username \"{str(username)}\" already taken"
+#             return render_template('register.html', message=message)
 
-        # Make sure password >= 5 characters
-        if len(password) < 5:
-            message = "Password must be 5 characters minimum"
-            return render_template('register.html', message=message)
+#         # Make sure password >= 5 characters
+#         if len(password) < 5:
+#             message = "Password must be 5 characters minimum"
+#             return render_template('register.html', message=message)
 
-        # Make sure password matches confirm password
-        if password != confirm_password:
-            message = "The two password fields must be identical"
-            return render_template('register.html', message=message)
+#         # Make sure password matches confirm password
+#         if password != confirm_password:
+#             message = "The two password fields must be identical"
+#             return render_template('register.html', message=message)
 
-        # Create cursor for database
-        curr = mysql.connection.cursor()
+#         # Create cursor for database
+#         curr = mysql.connection.cursor()
 
-        # Execute the insertion of the new user data
-        curr.execute("INSERT INTO users (username, name, email, password, create_time) VALUES (%s, %s, %s, %s, %s)", (username, name, email, encrypt_password, dToday))
+#         # Execute the insertion of the new user data
+#         curr.execute("INSERT INTO users (username, name, email, password, create_time) VALUES (%s, %s, %s, %s, %s)", (username, name, email, encrypt_password, dToday))
 
-        # Commit data to db
-        mysql.connection.commit()
+#         # Commit data to db
+#         mysql.connection.commit()
 
-        # Close connection
-        curr.close()
+#         # Close connection
+#         curr.close()
 
-        # Generate unique token for user
-        token = get_reset_token(username)
+#         # Generate unique token for user
+#         token = get_reset_token(username)
 
-        # Generate and send email through mailgun
-        send_simple_message("FridgeMate Registration Confirmation", username.capitalize(), str(email), token, "email.html", os.environ.get("API_BASE_URL"), os.environ.get("MAIL_API_KEY"),  os.environ.get("MAIL_ADDRESS"))
+#         # Generate and send email through mailgun
+#         send_simple_message("FridgeMate Registration Confirmation", username.capitalize(), str(email), token, "email.html", os.environ.get("API_BASE_URL"), os.environ.get("MAIL_API_KEY"),  os.environ.get("MAIL_ADDRESS"))
 
-        flash(f'You are now registered as a new user {username}, confirmation email has been sent', 'success')
-        return render_template('login.html')
-    else:
-        return render_template('register.html')
+#         flash(f'You are now registered as a new user {username}, confirmation email has been sent', 'success')
+#         return render_template('login.html')
+#     else:
+#         return render_template('register.html')
 
 # Confirm account verification email link is actually linked to an account
-@app.route('/verify_email/<token>', methods=['GET'])
-def verify_email(token):
+# @app.route('/verify_email/<token>', methods=['GET'])
+# def verify_email(token):
 
-    # Obtain username from token
-    user = verify_reset_token(token)
+#     # Obtain username from token
+#     user = verify_reset_token(token)
 
-    # Ensure confirmation link isn't expired
-    if user == None:
-        flash('The link has expired, you may send another confirmation link below', 'danger')
-        return redirect(url_for('resendConfirmation'))
+#     # Ensure confirmation link isn't expired
+#     if user == None:
+#         flash('The link has expired, you may send another confirmation link below', 'danger')
+#         return redirect(url_for('resendConfirmation'))
 
-    username = user['username']
+#     username = user['username']
 
-    # Initialize a cursor to verify account
-    cur = mysql.connection.cursor()
-    cur.execute('UPDATE users SET verified = %s WHERE username = %s', ['yes', username])
-    mysql.connection.commit()
-    cur.close()
+#     # Initialize a cursor to verify account
+#     cur = mysql.connection.cursor()
+#     cur.execute('UPDATE users SET verified = %s WHERE username = %s', ['yes', username])
+#     mysql.connection.commit()
+#     cur.close()
 
-    flash('Account verified', 'success')
-    return redirect(url_for('login'))
+#     flash('Account verified', 'success')
+#     return redirect(url_for('login'))
 
-@app.route('/resendConfirmation', methods=['POST', 'GET'])
-def resendConfirmation():
-    if request.method == 'POST':
+# @app.route('/resendConfirmation', methods=['POST', 'GET'])
+# def resendConfirmation():
+#     if request.method == 'POST':
 
-        # Verify email was entered correctly
-        email = str(request.form['email'])
-        if not email:
-            error = 'Email field cannot be blank'
-            return render_template('resendConfirmation.html', error=error)
+#         # Verify email was entered correctly
+#         email = str(request.form['email'])
+#         if not email:
+#             error = 'Email field cannot be blank'
+#             return render_template('resendConfirmation.html', error=error)
 
-        # Verify email has been registered in database
-        cur = mysql.connection.cursor()
-        found = cur.execute('SELECT * FROM users WHERE email = %s', [email])
-        if found == 0:
-            error='Email isn\t registered to an account (Go to \'Menu\'->\'Register\')'
-            return render_template('resendConfirmation.html', error=error)
+#         # Verify email has been registered in database
+#         cur = mysql.connection.cursor()
+#         found = cur.execute('SELECT * FROM users WHERE email = %s', [email])
+#         if found == 0:
+#             error='Email isn\t registered to an account (Go to \'Menu\'->\'Register\')'
+#             return render_template('resendConfirmation.html', error=error)
 
-        user = cur.fetchone()
-        username = user['username']
-        # Generate unique token for user
-        token = get_reset_token(username)
+#         user = cur.fetchone()
+#         username = user['username']
+#         # Generate unique token for user
+#         token = get_reset_token(username)
 
-        # Registration confirmation email
-        send_simple_message('FridgeMate Registration Confirmation', username, email, token, 'email.html', os.environ.get("API_BASE_URL"), os.environ.get("MAIL_API_KEY"),  os.environ.get("MAIL_ADDRESS"))
+#         # Registration confirmation email
+#         send_simple_message('FridgeMate Registration Confirmation', username, email, token, 'email.html', os.environ.get("API_BASE_URL"), os.environ.get("MAIL_API_KEY"),  os.environ.get("MAIL_ADDRESS"))
 
-        flash(f'A new confirmation email has been sent out to {email}', 'success')
-        return redirect(url_for('login'))
-    else:
-        return render_template('resendConfirmation.html')
+#         flash(f'A new confirmation email has been sent out to {email}', 'success')
+#         return redirect(url_for('login'))
+#     else:
+#         return render_template('resendConfirmation.html')
 
-# Log-in
-@app.route('/login', methods=['POST', 'GET'])
-def login():
+# # Log-in
+# @app.route('/login', methods=['POST', 'GET'])
+# def login():
 
-    if request.method == 'POST':
+#     if request.method == 'POST':
 
-        # Get form data
-        username = request.form['username']
-        password_entered = request.form['password']
+#         # Get form data
+#         username = request.form['username']
+#         password_entered = request.form['password']
 
-        #Before continuing, verify inputs aren't blank
-        if not username:
-            error = "Username cannot be left blank"
-            return render_template('/login.html', error=error)
+#         #Before continuing, verify inputs aren't blank
+#         if not username:
+#             error = "Username cannot be left blank"
+#             return render_template('/login.html', error=error)
 
-        if not password_entered:
-            error = "Password field cannot be blank"
-            return render_template('/login.html', error=error)
+#         if not password_entered:
+#             error = "Password field cannot be blank"
+#             return render_template('/login.html', error=error)
 
-        # First letter isn't case sensitive so must check lowercase and uppercase first letter
-        username_un_capitalize = username[0].lower() + username[1:]
+#         # First letter isn't case sensitive so must check lowercase and uppercase first letter
+#         username_un_capitalize = username[0].lower() + username[1:]
 
-        # Create cursor for database
-        cur = mysql.connection.cursor()
+#         # Create cursor for database
+#         cur = mysql.connection.cursor()
 
-        # Verify that the username entered in the form exists in the database
-        user_found = cur.execute('SELECT * FROM users WHERE username = %s OR username = %s OR username = %s', [username, username.capitalize(), username_un_capitalize])
+#         # Verify that the username entered in the form exists in the database
+#         user_found = cur.execute('SELECT * FROM users WHERE username = %s OR username = %s OR username = %s', [username, username.capitalize(), username_un_capitalize])
 
-        if user_found > 0:
-            data = cur.fetchone()
-            password = data['password']
+#         if user_found > 0:
+#             data = cur.fetchone()
+#             password = data['password']
 
-            if data['verified'] == 'no':
-                error = 'Account locked until email is verified.'
-                return render_template('login.html', error=error)
+#             if data['verified'] == 'no':
+#                 error = 'Account locked until email is verified.'
+#                 return render_template('login.html', error=error)
 
-            if sha256_crypt.verify(password_entered, password):
-                session['logged_in'] = True
-                session['id'] = data['id']
-                session['username'] = data['username']
-                flash(f'Welcome {username.capitalize()}', 'success')
-                cur.close()
-                return redirect(url_for('fridge_view'))
-            else:
-                cur.close()
-                error = 'Incorrect password'
-                return render_template('login.html', error=error)
+#             if sha256_crypt.verify(password_entered, password):
+#                 session['logged_in'] = True
+#                 session['id'] = data['id']
+#                 session['username'] = data['username']
+#                 flash(f'Welcome {username.capitalize()}', 'success')
+#                 cur.close()
+#                 return redirect(url_for('fridge_view'))
+#             else:
+#                 cur.close()
+#                 error = 'Incorrect password'
+#                 return render_template('login.html', error=error)
 
-        else:
-            error = 'Username does not exist'
-            return render_template('login.html', error=error)
+#         else:
+#             error = 'Username does not exist'
+#             return render_template('login.html', error=error)
 
-    else:
-        return render_template('login.html')
+#     else:
+#         return render_template('login.html')
 
-# Generate reset token for forgotten passwords and registeration
-def get_reset_token(user, expires=5000):
-    return jwt.encode({'reset_password': user, 'exp': time() + expires}, key=str(os.environ.get('SECRET_KEY')), algorithm="HS256")
+# # Generate reset token for forgotten passwords and registeration
+# def get_reset_token(user, expires=5000):
+#     return jwt.encode({'reset_password': user, 'exp': time() + expires}, key=str(os.environ.get('SECRET_KEY')), algorithm="HS256")
 
-# Verify reset token from user clicking link in email
-def verify_reset_token(token):
-    try:
-        username = jwt.decode(token, key=str(os.environ.get('SECRET_KEY')), algorithms="HS256")['reset_password']
-    except Exception as e:
-        print(e)
-        print('token has expired')
-        return None
-    cur = mysql.connection.cursor()
-    cur.execute('SELECT * FROM users WHERE username = %s', [username])
-    user = cur.fetchone()
-    cur.close()
-    return user
+# # Verify reset token from user clicking link in email
+# def verify_reset_token(token):
+#     try:
+#         username = jwt.decode(token, key=str(os.environ.get('SECRET_KEY')), algorithms="HS256")['reset_password']
+#     except Exception as e:
+#         print(e)
+#         print('token has expired')
+#         return None
+#     cur = mysql.connection.cursor()
+#     cur.execute('SELECT * FROM users WHERE username = %s', [username])
+#     user = cur.fetchone()
+#     cur.close()
+#     return user
 
-# Forgot Password
-@app.route('/forgotPass', methods=['POST', 'GET'])
-def forgotPass():
-    if request.method == 'POST':
+# # Forgot Password
+# @app.route('/forgotPass', methods=['POST', 'GET'])
+# def forgotPass():
+#     if request.method == 'POST':
 
-        # Create variable for email
-        email = request.form['email']
+#         # Create variable for email
+#         email = request.form['email']
 
-        # Verify email is actually associated with an account
-        cur = mysql.connection.cursor()
-        check = cur.execute('SELECT * FROM users WHERE email = %s', [email])
-        if check < 1:
-            cur.close()
-            error = 'Email isn\'t associated with an account'
-            return render_template('password_reset.html', error=error)
-        else:
-            profile = cur.fetchone()
-            cur.close()
+#         # Verify email is actually associated with an account
+#         cur = mysql.connection.cursor()
+#         check = cur.execute('SELECT * FROM users WHERE email = %s', [email])
+#         if check < 1:
+#             cur.close()
+#             error = 'Email isn\'t associated with an account'
+#             return render_template('password_reset.html', error=error)
+#         else:
+#             profile = cur.fetchone()
+#             cur.close()
 
-            username = profile['username']
+#             username = profile['username']
 
-            # Generate unique token for user
-            token = get_reset_token(username)
+#             # Generate unique token for user
+#             token = get_reset_token(username)
 
-            # Generate forgot password email
-            send_simple_message("Password Reset Confirmation", username.capitalize(), str(email), token, "password_reset_email.html", os.environ.get("API_BASE_URL"), os.environ.get("MAIL_API_KEY"),  os.environ.get("MAIL_ADDRESS"))
-            flash(f"Password reset link has been sent to {str(email)}", "success")
-            return redirect(url_for('login'))
+#             # Generate forgot password email
+#             send_simple_message("Password Reset Confirmation", username.capitalize(), str(email), token, "password_reset_email.html", os.environ.get("API_BASE_URL"), os.environ.get("MAIL_API_KEY"),  os.environ.get("MAIL_ADDRESS"))
+#             flash(f"Password reset link has been sent to {str(email)}", "success")
+#             return redirect(url_for('login'))
 
-    else:
-        return render_template('password_reset.html')
+#     else:
+#         return render_template('password_reset.html')
 
-# Reset password
-@app.route('/reset_password/<token>', methods=['POST', 'GET'])
-def reset_password(token):
-    if request.method == 'POST':
-        user = verify_reset_token(token)
-        username = user['username']
-        newPass = request.form['newPass']
-        confirmPass = request.form['confirmNewPass']
+# # Reset password
+# @app.route('/reset_password/<token>', methods=['POST', 'GET'])
+# def reset_password(token):
+#     if request.method == 'POST':
+#         user = verify_reset_token(token)
+#         username = user['username']
+#         newPass = request.form['newPass']
+#         confirmPass = request.form['confirmNewPass']
 
-        if not newPass or not confirmPass:
-            error='Password fields cannot be blank.'
-            return render_template('password_change.html', error=error, token=token)
+#         if not newPass or not confirmPass:
+#             error='Password fields cannot be blank.'
+#             return render_template('password_change.html', error=error, token=token)
 
-        if newPass != confirmPass:
-            error = 'Password fields must match.'
-            return render_template('password_change.html', error=error, token=token)
+#         if newPass != confirmPass:
+#             error = 'Password fields must match.'
+#             return render_template('password_change.html', error=error, token=token)
 
-        password = sha256_crypt.hash(str(newPass))
-        cur = mysql.connection.cursor()
-        cur.execute('UPDATE users SET password = %s WHERE username = %s', [password, username])
-        mysql.connection.commit()
-        cur.close()
-        flash('Password updated successfully', 'success')
-        return render_template('login.html')
-    else:
-        user = verify_reset_token(token)
-        if not user:
-            flash('No user found')
-            return redirect(url_for('login'))
+#         password = sha256_crypt.hash(str(newPass))
+#         cur = mysql.connection.cursor()
+#         cur.execute('UPDATE users SET password = %s WHERE username = %s', [password, username])
+#         mysql.connection.commit()
+#         cur.close()
+#         flash('Password updated successfully', 'success')
+#         return render_template('login.html')
+#     else:
+#         user = verify_reset_token(token)
+#         if not user:
+#             flash('No user found')
+#             return redirect(url_for('login'))
 
-        return render_template('password_change.html', token=token)
+#         return render_template('password_change.html', token=token)
 
 # User profile
 @app.route('/profile', methods=['POST', 'GET'])
@@ -782,6 +786,8 @@ def deleteEntries():
     cur.close()
 
     return message
+
+app.register_blueprint(userAuth)
 
 if __name__ == '__main__':
     app.run(debug=True)
